% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_hyp.R
\name{test_hyp}
\alias{test_hyp}
\title{Testing Informed Hypotheses}
\usage{
test_hyp(object, hyp, mcrep = 1e+06)
}
\arguments{
\item{object}{A regression model object fit using the \code{lm} function.}

\item{hyp}{A string specifying hypotheses to be tested using the variable
names of the \code{lm} object.}

\item{mcrep}{Integer specifying the number of iterations if no analytical
solutions is possible. This is rare and only the case if the rank of the
constraint matrix is less than its number of rows.}
}
\description{
Test hypotheses about continous predictors in an \code{\link{lm}}-object.
}
\details{
This function is based on a method by Mulder (2014), a modification of the
fractional Bayes factor approach. In essence, it uses a number of observations
equal to the number of predictors in the \code{lm} model to construct a
minimally informative prior, and the remainder of the observations are then
used to test the hypotheses.

The function requires that relevant variables have been standardized before
fitting the model with \code{lm}. This is done simply by substracting the mean
of a variable from each observation and dividing by the standard deviation. A
simple option for achieving this is to use the \code{\link{scale}} function.

Multiple hypotheses can be specified at the same time by separating them with
a semicolon. It is advisable to only specify competing hypotheses in this way,
that is, hypotheses regarding the same variables, e.g., \dQuote{X1 > 0; X1 <
0; X1 = 0}. If specifying multiple hypotheses and comparing against a value it
is currently only possible to compare against the same value, e.g., \dQuote{X1
= 0; X1 = 2} is not functional input. This is because the prior is centered
around the input value (or zero if no input value), which is not possible in
the case of several input values.

Parentheses can be used to compare multiple variables with the same variable
or value. For example, \dQuote{(X1, X2) > X3} is read as \dQuote{X1 > X3 and
X2 > X3}.

For each specified hypothesis the posterior probability is output. If the
hypotheses are not mutually exhaustive this includes the posterior probability
of the complement to the input hypotheses. For example, inputting \dQuote{X1 >
0; X1 < 0} gives posterior probabilities for only for these hypotheses,
whereas inputting \dQuote{(X1, X2) > 0} gives posterior probabilities for
\dQuote{(X1, X2) > 0} and \dQuote{not (X1, X2) > 0}.

By saving the test as an object it is also possible to access the
\code{BF_matrix} which compares the hypotheses directly against each other
(see examples). This matrix divides the column hypothesis by each row
hypothesis and can be interpreted as \dQuote{given the data, [column
hypothesis] is [value] times as likely as [row hypothesis]}.
}
\section{References}{
 Mulder, J. (2014). Prior adjusted default Bayes factors
 for testing (in) equality constrained hypotheses. Computational Statistics &
 Data Analysis, 71, 448-463.
}

\examples{
###Standardize variables and fit the linear model
dt <- as.data.frame(scale(mtcars[, c(1, 3:4, 6)]))
fit <- lm(mpg ~ disp + hp + wt, data = dt)

###Define hypotheses based on theory and test them
hyp <- "(wt, hp) > disp > 0; (wt, hp) > disp = 0"
res <- test_hyp(fit, hyp)
res

###Examine output that compares hypotheses directly with Bayes factors
res$BF_matrix

}
